pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Calculates the differences between two files of json or yaml formats."
authors = ["Lera <leilanimoone@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]
repository = "https://github.com/leilanimoone/python-project-50"
classifiers = ["Topic :: Software Development :: Libraries :: Python Modules"]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
